--- dotnet/src/Extensions/TemplateEngine.Basic/TemplateEngine.Basic.csproj
@@ -1,25 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <!-- THIS PROPERTY GROUP MUST COME FIRST -->
-    <AssemblyName>Microsoft.SemanticKernel.TemplateEngine.Basic</AssemblyName>
-    <RootNamespace>Microsoft.SemanticKernel.TemplateEngine.Basic</RootNamespace>
-    <TargetFramework>netstandard2.0</TargetFramework>
-  </PropertyGroup>
-
-  <Import Project="$(RepoRoot)/dotnet/nuget/nuget-package.props" />
-
-  <PropertyGroup>
-    <!-- NuGet Package Settings -->
-    <Title>Semantic Kernel - Basic Prompt Template Engine</Title>
-    <Description>Semantic Kernel Basic Prompt Template Engine</Description>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <InternalsVisibleTo Include="SemanticKernel.Extensions.UnitTests" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\..\SemanticKernel.Core\SemanticKernel.Core.csproj" />
-  </ItemGroup>
-</Project>
